requirements.txt
streamlit
tensorflow
numpy
Pillow
opencv-python
scikit-image
scipy
scikit-learn